Fix Cas Hours Float Bugs
////////

+++ Assets/Resources/Scripts/UserManager.cs
@@ -8,7 +8,7 @@ public class UserManager {
 	public static int id;
 	public static int level;
 	public static int exp;
-	public static int casHours;
+	public static double casHours;
 	public static double hourRewarded;
 	public static DateTime rewardDate;
 	public static bool isAdmin;
